Refactor network tags and timekeeping organization
- Remove NetworkTags; migrate all usage to TaskNaming.Tags.
- Move TimeSynchronizer and TimingWheel to new namespaces.
- Move ConnectionLimiter to Internal.Limiters and make internal.
- Remove listener time sync control and related reporting.
- Update all worker/task scheduling to use TaskNaming.Tags.
- Clean up usings, namespaces, and project folder structure.
