+++ src/Nalix.Network/Nalix.Network.csproj
@@ -41,4 +41,8 @@ Built for real-time workloads, it emphasizes zero-allocation I/O, low-latency di
 
 		<None Include="$(MSBuildProjectDirectory)\..\..\build\bin\$(Configuration)\Nalix.Analyzers.CodeFixes.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Condition="Exists('$(MSBuildProjectDirectory)\..\..\build\bin\$(Configuration)\Nalix.Analyzers.CodeFixes.dll')" />
 	</ItemGroup>
+
+	<ItemGroup>
+	  <Folder Include="Timekeeping\" />
+	</ItemGroup>
 </Project>
