+++ src/Nalix.Network/Routing/PacketDispatchChannel.cs
@@ -21,7 +21,6 @@ using Nalix.Framework.Extensions;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Options;
 using Nalix.Framework.Tasks;
-using Nalix.Network.Internal.Constants;
 using Nalix.Network.Internal.Routing;
 
 namespace Nalix.Network.Routing;
@@ -130,14 +129,14 @@ public sealed class PacketDispatchChannel
 
             _workerHandle[i] = InstanceManager.Instance.GetOrCreateInstance<TaskManager>().ScheduleWorker(
                 name: $"{TaskNaming.Tags.Dispatch}.{TaskNaming.Tags.Process}.{i}",
-                group: $"{NetworkTags.Net}/{TaskNaming.Tags.Dispatch}",
+                group: $"{TaskNaming.Tags.Net}/{TaskNaming.Tags.Dispatch}",
                 work: async (ctx, ct) => await this.RunLoop(ctx, ct).ConfigureAwait(false),
                 options: new WorkerOptions
                 {
+                    RetainFor = TimeSpan.Zero,
+                    Tag = TaskNaming.Tags.Net,
                     IdType = SnowflakeType.System,
                     CancellationToken = linkedToken,
-                    RetainFor = TimeSpan.Zero,
-                    Tag = NetworkTags.Net
                 });
         }
